Move `editor` module to its own dir.
