+++ editor/editor.js
